Switch from toDataURL to getImageData
Because `Canvas#toDataURL` generates a compressed PNG image, it
generates different results on different platforms. Thus, it can't be
used to generate snapshots that are supposed to work across different
platforms.

`CanvasRenderingContext2D.getImageData` on the other hand, returns raw
image data, so it returns exactly the same data regardless of the
platform. Size of returned data is not an issue, because the result is
hashed anyway.